alsa-firmware: remove nocross, add ignore_elf_files
alsa-firmware's only C files are internal generation utilities;
use the host cc to allow it to be cross.

There is an elf file that is required for one of the pieces of
firmware which needs to be located in /usr/share; add it to
ignore_elf_files. This will be required for the nocross removal.